Partially split CommonConstructors.cpp into few separate files
